Feature/NSP-44 app home page with mix contents l

+++ lib/features/home/data/model/resume/resume_model.dart
@@ -1,47 +1,3 @@
-// To parse this JSON data, do
-//
-//     final resume = resumeFromJson(jsonString);
-
-// class Resume {
-//     Resume({
-//         this.type,
-//         this.data,
-//     });
-
-//     String type;
-//     ResumeData data;
-
-//     factory Resume.fromJson(Map<String, dynamic> json) => Resume(
-//         type: json["type"],
-//         data: ResumeData.fromJson(json["data"]),
-//     );
-
-//     Map<String, dynamic> toJson() => {
-//         "type": type,
-//         "data": data.toJson(),
-//     };
-// }
-
-// class ResumeData {
-//     ResumeData({
-//         this.status,
-//         this.data,
-//     });
-
-//     bool status;
-//     DataData data;
-
-//     factory ResumeData.fromJson(Map<String, dynamic> json) => ResumeData(
-//         status: json["status"],
-//         data: DataData.fromJson(json["data"]),
-//     );
-
-//     Map<String, dynamic> toJson() => {
-//         "status": status,
-//         "data": data.toJson(),
-//     };
-// }
-
 import 'package:flutter/foundation.dart';
 import 'package:wallet_app/features/home/data/model/resume/personal_info_model.dart';
 import 'package:wallet_app/features/home/data/model/resume/qualification_history_model.dart';
@@ -54,13 +10,13 @@ import 'academic_history_model.dart';
 class ResumeDataModel extends ResumeData {
   const ResumeDataModel({
     @required PersonalInfoModel personalInfo,
-    @required double personalInfoCompletionRate,
+    @required int personalInfoCompletionRate,
     @required List<AcademicHistoryModel> academicHistory,
-    @required double academicsCompletionRate,
+    @required int academicsCompletionRate,
     @required List<WorkHistoryModel> workHistory,
-    @required double worksCompletionRate,
+    @required int worksCompletionRate,
     @required List<QualificationHistoryModel> qualificationHistory,
-    @required double qualificationCompletionRate,
+    @required int qualificationCompletionRate,
     @required ResumeOptionsModel options,
   }) : super(
           personalInfo: personalInfo,
@@ -79,23 +35,38 @@ class ResumeDataModel extends ResumeData {
         personalInfo: PersonalInfoModel.fromJson(
             json["personal_info"] as Map<String, dynamic>),
         personalInfoCompletionRate:
-            json["personal_info_completion_rate"] as double,
+            json["personal_info_completion_rate"] as int,
         academicHistory: List<AcademicHistoryModel>.from(
             (json["academic_history"] as Iterable).map((x) =>
                 AcademicHistoryModel.fromJson(x as Map<String, dynamic>))),
-        academicsCompletionRate: json["academics_completion_rate"] as double,
+        academicsCompletionRate: json["academics_completion_rate"] as int,
         workHistory: List<WorkHistoryModel>.from((json["work_history"]
                 as Iterable)
             .map((x) => WorkHistoryModel.fromJson(x as Map<String, dynamic>))),
-        worksCompletionRate: json["works_completion_rate"] as double,
+        worksCompletionRate: json["works_completion_rate"] as int,
         qualificationHistory: List<QualificationHistoryModel>.from(
             (json["qualification_history"] as Iterable).map((x) =>
                 QualificationHistoryModel.fromJson(x as Map<String, dynamic>))),
         qualificationCompletionRate:
-            json["qualification_completion_rate"] as double,
+            json["qualification_completion_rate"] as int,
         options: ResumeOptionsModel.fromJson(
             json["options"] as Map<String, dynamic>),
       );
+
+  // factory ResumeDataModel.fromJson(Map<String, dynamic> json) =>
+  //     ResumeDataModel(
+  //       personalInfo: PersonalInfoModel.fromJson(
+  //           json["personal_info"] as Map<String, dynamic>),
+  //       personalInfoCompletionRate:
+  //           (json["personal_info_completion_rate"] as int).toDouble(),
+  //       academicHistory: null,
+  //       academicsCompletionRate: null,
+  //       workHistory: null,
+  //       worksCompletionRate: null,
+  //       qualificationHistory: null,
+  //       qualificationCompletionRate: null,
+  //       options: null,
+  //     );
 
   Map<String, dynamic> toJson() => {
         "personal_info": (personalInfo as PersonalInfoModel).toJson(),
